+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
     "cSpell.words": [
         "Allnews",
+        "ARGB",
         "firstname",
         "lastname",
         "LTRB",
