+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:provider/provider.dart';
 import 'package:snews/src/feature/presentation/bloc/all_news_bloc.dart';
-import 'package:snews/src/feature/presentation/screen/main_screen.dart';
+import 'package:snews/src/feature/presentation/screens/main_screen.dart';
 
 void main() {
   runApp(const MainApp());
