+++ lib/src/feature/data/data_api_impl.dart
@@ -12,7 +12,7 @@ class ApiDataImpl implements IDataApi {
   Future<AllNewsEntity> fetchAllNews() async {
     final response = await _client!.get(
       Uri.parse(
-          'https://api.nytimes.com/svc/search/v2/articlesearch.json?api-key=7G6FwHXsXDGLikbSgrCwcOHABUWIeTrc'),
+          'https://api.nytimes.com/svc/topstories/v2/business.json?api-key=7G6FwHXsXDGLikbSgrCwcOHABUWIeTrc'),
       headers: {
         'Content-Type': 'application/json',
       },
